Create a README.md file
"For a given period (typically half an hour) the sensible
heat flux at any height is the net exchange of heat conducted
by all the ramps formed during this period. ....... A and
t (tau) are respectively the mean ramp amplitude
and the inverse ramp frequency. r is the time (seconds) when the maximum ratio of  3rd moment to time (seconds) is observed.